Transparent background by default
A transparent root element is useful for Preloaders and 3D Contexts.

Reasoning that a default value of 0 transparency is fine for all cases because backgroundColor value is determined from reading the containing div's style tag (line 258).  So if it's been set then the CSS already exists to set it as that backgroundColor (or backgroundImage) by following the default CSS flow.  If, on the other hand, the backgroundColor has not been set then the default CSS behavior of transparent should be followed.

The drawRect is still needed for the hitTest events as the original comment states.